+++ src/components/RBarcode/index.ts
@@ -1,0 +1,8 @@
+import RBarcode from './src/Barcode'
+import barcodeProps from './src/props'
+
+import type { ExtractPublicPropTypes } from 'vue'
+
+export type BarcodeProps = ExtractPublicPropTypes<typeof barcodeProps>
+
+export { RBarcode, barcodeProps }