+++ src/components/index.ts
@@ -10,6 +10,7 @@ export * from './RTable'
 export * from './RTransitionComponent'
 export * from './RForm'
 export * from './RSegment'
+export * from './RBarcode'
 
 // 导出自定义组件类型
 export type * from './RChart/src/types'
@@ -21,3 +22,4 @@ export type * from './RTransitionComponent/src/types'
 export type * from './RForm/src/types'
 export type * from './RModal/src/types'
 export type * from './RSegment/src/types'
+export type * from './RBarcode/src/types'